Face-Detection/requirements.txt
azure-cognitiveservices-vision-face
pillow>=9.0.0
matplotlib==3.5.1
numpy>=1.21.0
opencv-python-headless